# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Luxury company platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Bad request

  # USERS (admin only)
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      parameters:
        - name: featured
          in: query
          schema:
            type: boolean
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Projects list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # CASE STUDIES
  /case-studies:
    get:
      operationId: listCaseStudies
      tags: [caseStudies]
      summary: List all case studies
      parameters:
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Case studies list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CaseStudy"
    post:
      operationId: createCaseStudy
      tags: [caseStudies]
      summary: Create a case study
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseStudyInput"
      responses:
        "201":
          description: Created case study
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseStudy"
  /case-studies/{id}:
    get:
      operationId: getCaseStudy
      tags: [caseStudies]
      summary: Get a case study by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Case study
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseStudy"
        "404":
          description: Not found
    patch:
      operationId: updateCaseStudy
      tags: [caseStudies]
      summary: Update a case study
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseStudyUpdate"
      responses:
        "200":
          description: Updated case study
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseStudy"
    delete:
      operationId: deleteCaseStudy
      tags: [caseStudies]
      summary: Delete a case study
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # SERVICES
  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List all services
      responses:
        "200":
          description: Services list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [services]
      summary: Create a service
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Created service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{id}:
    get:
      operationId: getService
      tags: [services]
      summary: Get a service by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "404":
          description: Not found
    patch:
      operationId: updateService
      tags: [services]
      summary: Update a service
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceUpdate"
      responses:
        "200":
          description: Updated service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete a service
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # COMPANY
...
  /media:
    get:
      operationId: listMedia
      tags: [media]
      summary: List all media files
      parameters:
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Media list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Media"
  /media/{id}:
    get:
      operationId: getMedia
      tags: [media]
      summary: Get a media file by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Media
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Media"
    patch:
      operationId: updateMedia
      tags: [media]
      summary: Update media metadata
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MediaUpdate"
      responses:
        "200":
          description: Updated media
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Media"
    delete:
      operationId: deleteMedia
      tags: [media]
      summary: Delete a media file
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # SETTINGS
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # AUTH
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 6

    # USERS
    User:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
...
    Project:
      type: object
      required: [id, title, slug, description, status, featured, sortOrder, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        content:
          type: ["string", "null"]
        client:
          type: ["string", "null"]
        year:
          type: ["integer", "null"]
        category:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]
        coverImageUrl:
          type: ["string", "null"]
        imageUrls:
          type: ["string", "null"]
        status:
          type: string
        featured:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
    ProjectInput:
      type: object
      required: [title, slug, description, status]
      properties:
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        content:
          type: string
        client:
          type: string
        year:
          type: integer
        category:
          type: string
        tags:
          type: string
        coverImageUrl:
          type: string
        imageUrls:
          type: string
        status:
          type: string
        featured:
          type: boolean
        sortOrder:
          type: integer
    ProjectUpdate:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        content:
          type: string
        client:
          type: string
        year:
          type: integer
        category:
          type: string
        tags:
          type: string
        coverImageUrl:
          type: string
        imageUrls:
          type: string
        status:
          type: string
        featured:
          type: boolean
        sortOrder:
          type: integer

    # CASE STUDIES
    CaseStudy:
      type: object
      required: [id, title, slug, summary, challenge, solution, outcome, featured, sortOrder, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        summary:
          type: string
        challenge:
          type: string
        solution:
          type: string
        outcome:
          type: string
        client:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        duration:
          type: ["string", "null"]
        coverImageUrl:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]
        metrics:
          type: ["string", "null"]
        featured:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
    CaseStudyInput:
      type: object
      required: [title, slug, summary, challenge, solution, outcome]
      properties:
        title:
          type: string
        slug:
          type: string
        summary:
          type: string
        challenge:
          type: string
        solution:
          type: string
        outcome:
          type: string
        client:
          type: string
        industry:
          type: string
        duration:
          type: string
        coverImageUrl:
          type: string
        tags:
          type: string
        metrics:
          type: string
        featured:
          type: boolean
        sortOrder:
          type: integer
    CaseStudyUpdate:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        summary:
          type: string
        challenge:
          type: string
        solution:
          type: string
        outcome:
          type: string
        client:
          type: string
        industry:
          type: string
        duration:
          type: string
        coverImageUrl:
          type: string
        tags:
          type: string
        metrics:
          type: string
        featured:
          type: boolean
        sortOrder:
          type: integer

    # SERVICES
    Service:
      type: object
      required: [id, title, slug, description, sortOrder, active, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        content:
          type: ["string", "null"]
        iconName:
          type: ["string", "null"]
        coverImageUrl:
          type: ["string", "null"]
        features:
          type: ["string", "null"]
        sortOrder:
          type: integer
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
    ServiceInput:
      type: object
      required: [title, slug, description]
      properties:
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        content:
          type: string
        iconName:
          type: string
        coverImageUrl:
          type: string
        features:
          type: string
        sortOrder:
          type: integer
        active:
          type: boolean
    ServiceUpdate:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        content:
          type: string
        iconName:
          type: string
        coverImageUrl:
          type: string
        features:
          type: string
        sortOrder:
          type: integer
        active:
          type: boolean

    # COMPANY
...
    Media:
      type: object
      required: [id, filename, originalName, mimeType, size, url, createdAt]
      properties:
        id:
          type: integer
        filename:
          type: string
        originalName:
          type: string
        mimeType:
          type: string
        size:
          type: integer
        url:
          type: string
        alt:
          type: ["string", "null"]
        caption:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    MediaUpdate:
      type: object
      properties:
        alt:
          type: string
        caption:
          type: string

    # SETTINGS
...