# lib/api-zod/src/generated/types/media.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Luxury company platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Media {
  id: number;
  filename: string;
  originalName: string;
  mimeType: string;
  size: number;
  url: string;
  /** @nullable */
  alt?: string | null;
  /** @nullable */
  caption?: string | null;
  createdAt: Date;
}