# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Luxury company platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: number;
  title: string;
  slug: string;
  description: string;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  iconName?: string | null;
  /** @nullable */
  coverImageUrl?: string | null;
  /** @nullable */
  features?: string | null;
  sortOrder: number;
  active: boolean;
  createdAt: Date;
  /** @nullable */
  updatedAt?: Date | null;
}