# lib/api-zod/src/generated/types/serviceUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Luxury company platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ServiceUpdate {
  title?: string;
  slug?: string;
  description?: string;
  content?: string;
  iconName?: string;
  coverImageUrl?: string;
  features?: string;
  sortOrder?: number;
  active?: boolean;
}